requirements.txt
Django==2.2.8
djangorestframework==3.10.3
entrypoints==0.3
flake8==3.7.8
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.3
sqlparse==0.3.0
psycopg2==2.8.3
Pillow>=6.2.0
django-cors-headers